[iris] Fold classify_log_level into log_keys instead of its own module
A 4-line classifier did not warrant a standalone module. log_keys already
owns the iris-domain to finelog mappings (keys, match scopes) and is
imported by the worker; the level classifier is a direct sibling.
