+++ lib/iris/src/iris/cluster/log_keys.py
@@ -1,18 +1,47 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Iris-domain helpers that produce opaque string keys for the finelog store.
+"""Iris-domain helpers that map iris values onto the finelog store.
 
-Keys are plain strings on the finelog side; this module owns the mapping from
-iris-domain values (`JobName`, `TaskAttempt`) to those strings.
+This module owns the mappings between iris-domain values (`JobName`,
+`TaskAttempt`, capture streams) and the finelog wire types: the opaque string
+keys that identify log streams, and the `LogLevel` assigned to each line.
 """
 
 from finelog.rpc import logging_pb2
+from finelog.types import str_to_log_level
+from rigging.log_setup import parse_log_level
 
 from iris.cluster.types import JobName, TaskAttempt
 
 CONTROLLER_LOG_KEY = "/system/controller"
 _WORKER_LOG_PREFIX = "/system/worker/"
+
+# Default level per capture stream when a line carries no parseable prefix.
+# "error" is the synthetic source iris uses for injected failure lines (OOM
+# kills, infrastructure errors). Streams not listed here (e.g. "build") fall
+# back to UNKNOWN, which stays visible under every min_level filter.
+_STREAM_DEFAULT_LEVEL = {
+    "stdout": logging_pb2.LOG_LEVEL_INFO,
+    "stderr": logging_pb2.LOG_LEVEL_ERROR,
+    "error": logging_pb2.LOG_LEVEL_ERROR,
+}
+
+
+def classify_log_level(source: str, data: str) -> int:
+    """Assign a finelog ``LogLevel`` to a captured task log line.
+
+    A parseable glog-style level prefix in ``data`` always wins, so a prefixed
+    ``INFO`` line on ``stderr`` is classified ``INFO``, not ``ERROR``. Otherwise
+    the level defaults from ``source`` (the capture stream): ``stdout`` is
+    informational, ``stderr`` and iris's injected failure lines are errors. This
+    keeps mundane stdout out of the ``min_level``-filtered error view, where an
+    ``UNKNOWN`` line would otherwise pass through every filter.
+    """
+    parsed = str_to_log_level(parse_log_level(data))
+    if parsed != logging_pb2.LOG_LEVEL_UNKNOWN:
+        return parsed
+    return _STREAM_DEFAULT_LEVEL.get(source, logging_pb2.LOG_LEVEL_UNKNOWN)
 
 
 def worker_log_key(worker_id: str) -> str:
